+++ src/lw_common/ui/search/search_form.Designer.cs
@@ -95,16 +95,19 @@
             this.combo = new System.Windows.Forms.ComboBox();
             this.markAsNewEntry = new System.Windows.Forms.CheckBox();
             this.allColumns = new System.Windows.Forms.CheckBox();
+            this.toFilter = new System.Windows.Forms.Button();
+            this.filterHelp = new System.Windows.Forms.LinkLabel();
+            this.negate = new System.Windows.Forms.CheckBox();
             ((System.ComponentModel.ISupportInitialize)(this.result)).BeginInit();
             this.SuspendLayout();
             // 
             // label1
             // 
             this.label1.AutoSize = true;
-            this.label1.Location = new System.Drawing.Point(13, 20);
-            this.label1.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
+            this.label1.Location = new System.Drawing.Point(26, 40);
+            this.label1.Margin = new System.Windows.Forms.Padding(8, 0, 8, 0);
             this.label1.Name = "label1";
-            this.label1.Size = new System.Drawing.Size(35, 17);
+            this.label1.Size = new System.Drawing.Size(67, 31);
             this.label1.TabIndex = 0;
             this.label1.Text = "Text";
             // 
@@ -114,9 +117,10 @@
             this.mark.AutoSize = true;
             this.mark.Checked = true;
             this.mark.CheckState = System.Windows.Forms.CheckState.Checked;
-            this.mark.Location = new System.Drawing.Point(277, 110);
+            this.mark.Location = new System.Drawing.Point(575, 220);
+            this.mark.Margin = new System.Windows.Forms.Padding(6);
             this.mark.Name = "mark";
-            this.mark.Size = new System.Drawing.Size(202, 21);
+            this.mark.Size = new System.Drawing.Size(383, 35);
             this.mark.TabIndex = 2;
             this.mark.Text = "Mark found words with color";
             this.mark.UseVisualStyleBackColor = true;
@@ -126,9 +130,11 @@
             this.fg.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
             this.fg.BackColor = System.Drawing.SystemColors.Control;
             this.fg.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
-            this.fg.Location = new System.Drawing.Point(514, 110);
+            this.fg.Cursor = System.Windows.Forms.Cursors.Hand;
+            this.fg.Location = new System.Drawing.Point(1028, 220);
+            this.fg.Margin = new System.Windows.Forms.Padding(6, 0, 6, 0);
             this.fg.Name = "fg";
-            this.fg.Size = new System.Drawing.Size(28, 17);
+            this.fg.Size = new System.Drawing.Size(54, 32);
             this.fg.TabIndex = 3;
             this.fg.Click += new System.EventHandler(this.fg_Click);
             // 
@@ -136,9 +142,10 @@
             // 
             this.label3.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
             this.label3.AutoSize = true;
-            this.label3.Location = new System.Drawing.Point(484, 109);
+            this.label3.Location = new System.Drawing.Point(968, 218);
+            this.label3.Margin = new System.Windows.Forms.Padding(6, 0, 6, 0);
             this.label3.Name = "label3";
-            this.label3.Size = new System.Drawing.Size(24, 17);
+            this.label3.Size = new System.Drawing.Size(46, 31);
             this.label3.TabIndex = 4;
             this.label3.Text = "Fg";
             // 
@@ -146,9 +153,10 @@
             // 
             this.label4.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
             this.label4.AutoSize = true;
-            this.label4.Location = new System.Drawing.Point(548, 110);
+            this.label4.Location = new System.Drawing.Point(1096, 220);
+            this.label4.Margin = new System.Windows.Forms.Padding(6, 0, 6, 0);
             this.label4.Name = "label4";
-            this.label4.Size = new System.Drawing.Size(25, 17);
+            this.label4.Size = new System.Drawing.Size(47, 31);
             this.label4.TabIndex = 6;
             this.label4.Text = "Bg";
             // 
@@ -156,18 +164,21 @@
             // 
             this.bg.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
             this.bg.BorderStyle = System.Windows.Forms.BorderStyle.FixedSingle;
-            this.bg.Location = new System.Drawing.Point(578, 110);
+            this.bg.Cursor = System.Windows.Forms.Cursors.Hand;
+            this.bg.Location = new System.Drawing.Point(1156, 220);
+            this.bg.Margin = new System.Windows.Forms.Padding(6, 0, 6, 0);
             this.bg.Name = "bg";
-            this.bg.Size = new System.Drawing.Size(28, 18);
+            this.bg.Size = new System.Drawing.Size(54, 34);
             this.bg.TabIndex = 5;
             this.bg.Click += new System.EventHandler(this.bg_Click);
             // 
             // ok
             // 
             this.ok.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
-            this.ok.Location = new System.Drawing.Point(450, 137);
+            this.ok.Location = new System.Drawing.Point(900, 274);
+            this.ok.Margin = new System.Windows.Forms.Padding(6);
             this.ok.Name = "ok";
-            this.ok.Size = new System.Drawing.Size(75, 25);
+            this.ok.Size = new System.Drawing.Size(150, 50);
             this.ok.TabIndex = 7;
             this.ok.Text = "OK";
             this.ok.UseVisualStyleBackColor = true;
@@ -177,9 +188,10 @@
             // 
             this.cancel.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
             this.cancel.DialogResult = System.Windows.Forms.DialogResult.Cancel;
-            this.cancel.Location = new System.Drawing.Point(531, 137);
+            this.cancel.Location = new System.Drawing.Point(1062, 274);
+            this.cancel.Margin = new System.Windows.Forms.Padding(6);
             this.cancel.Name = "cancel";
-            this.cancel.Size = new System.Drawing.Size(75, 25);
+            this.cancel.Size = new System.Drawing.Size(150, 50);
             this.cancel.TabIndex = 8;
             this.cancel.Text = "Cancel";
             this.cancel.UseVisualStyleBackColor = true;
@@ -190,9 +202,10 @@
             this.radioAutoRecognize.AutoSize = true;
             this.radioAutoRecognize.Checked = true;
             this.radioAutoRecognize.Font = new System.Drawing.Font("Microsoft Sans Serif", 8F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.radioAutoRecognize.Location = new System.Drawing.Point(68, 65);
+            this.radioAutoRecognize.Location = new System.Drawing.Point(136, 130);
+            this.radioAutoRecognize.Margin = new System.Windows.Forms.Padding(6);
             this.radioAutoRecognize.Name = "radioAutoRecognize";
-            this.radioAutoRecognize.Size = new System.Drawing.Size(96, 17);
+            this.radioAutoRecognize.Size = new System.Drawing.Size(188, 30);
             this.radioAutoRecognize.TabIndex = 9;
             this.radioAutoRecognize.TabStop = true;
             this.radioAutoRecognize.Text = "Auto recognize";
@@ -203,9 +216,10 @@
             // 
             this.radioText.AutoSize = true;
             this.radioText.Font = new System.Drawing.Font("Microsoft Sans Serif", 8F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.radioText.Location = new System.Drawing.Point(358, 65);
+            this.radioText.Location = new System.Drawing.Point(716, 130);
+            this.radioText.Margin = new System.Windows.Forms.Padding(6);
             this.radioText.Name = "radioText";
-            this.radioText.Size = new System.Drawing.Size(46, 17);
+            this.radioText.Size = new System.Drawing.Size(84, 30);
             this.radioText.TabIndex = 10;
             this.radioText.Text = "Text";
             this.radioText.UseVisualStyleBackColor = true;
@@ -215,9 +229,10 @@
             // 
             this.radioRegex.AutoSize = true;
             this.radioRegex.Font = new System.Drawing.Font("Microsoft Sans Serif", 8F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.radioRegex.Location = new System.Drawing.Point(422, 65);
+            this.radioRegex.Location = new System.Drawing.Point(844, 130);
+            this.radioRegex.Margin = new System.Windows.Forms.Padding(6);
             this.radioRegex.Name = "radioRegex";
-            this.radioRegex.Size = new System.Drawing.Size(56, 17);
+            this.radioRegex.Size = new System.Drawing.Size(106, 30);
             this.radioRegex.TabIndex = 11;
             this.radioRegex.Text = "Regex";
             this.radioRegex.UseVisualStyleBackColor = true;
@@ -226,9 +241,10 @@
             // caseSensitive
             // 
             this.caseSensitive.AutoSize = true;
-            this.caseSensitive.Location = new System.Drawing.Point(241, 86);
+            this.caseSensitive.Location = new System.Drawing.Point(482, 172);
+            this.caseSensitive.Margin = new System.Windows.Forms.Padding(6);
             this.caseSensitive.Name = "caseSensitive";
-            this.caseSensitive.Size = new System.Drawing.Size(121, 21);
+            this.caseSensitive.Size = new System.Drawing.Size(230, 35);
             this.caseSensitive.TabIndex = 12;
             this.caseSensitive.Text = "Case-Sensitive";
             this.caseSensitive.UseVisualStyleBackColor = true;
@@ -236,9 +252,10 @@
             // fullWord
             // 
             this.fullWord.AutoSize = true;
-            this.fullWord.Location = new System.Drawing.Point(368, 86);
+            this.fullWord.Location = new System.Drawing.Point(736, 172);
+            this.fullWord.Margin = new System.Windows.Forms.Padding(6);
             this.fullWord.Name = "fullWord";
-            this.fullWord.Size = new System.Drawing.Size(87, 21);
+            this.fullWord.Size = new System.Drawing.Size(161, 35);
             this.fullWord.TabIndex = 13;
             this.fullWord.Text = "Full Word";
             this.fullWord.UseVisualStyleBackColor = true;
@@ -352,12 +369,13 @@
             this.olvColumn50,
             this.olvColumn51});
             this.result.FullRowSelect = true;
-            this.result.Location = new System.Drawing.Point(12, 165);
+            this.result.Location = new System.Drawing.Point(24, 398);
+            this.result.Margin = new System.Windows.Forms.Padding(6);
             this.result.MultiSelect = false;
             this.result.Name = "result";
             this.result.OwnerDraw = true;
             this.result.ShowGroups = false;
-            this.result.Size = new System.Drawing.Size(594, 182);
+            this.result.Size = new System.Drawing.Size(1184, 344);
             this.result.TabIndex = 14;
             this.result.UseCompatibleStateImageBehavior = false;
             this.result.View = System.Windows.Forms.View.Details;
@@ -571,19 +589,20 @@
             this.friendlyRegexName.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
             this.friendlyRegexName.Enabled = false;
-            this.friendlyRegexName.Location = new System.Drawing.Point(483, 19);
+            this.friendlyRegexName.Location = new System.Drawing.Point(966, 38);
+            this.friendlyRegexName.Margin = new System.Windows.Forms.Padding(6);
             this.friendlyRegexName.Name = "friendlyRegexName";
-            this.friendlyRegexName.Size = new System.Drawing.Size(129, 23);
+            this.friendlyRegexName.Size = new System.Drawing.Size(254, 38);
             this.friendlyRegexName.TabIndex = 1;
             // 
             // label2
             // 
             this.label2.AutoSize = true;
             this.label2.Font = new System.Drawing.Font("Microsoft Sans Serif", 7F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.label2.Location = new System.Drawing.Point(483, 3);
-            this.label2.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
+            this.label2.Location = new System.Drawing.Point(966, 6);
+            this.label2.Margin = new System.Windows.Forms.Padding(8, 0, 8, 0);
             this.label2.Name = "label2";
-            this.label2.Size = new System.Drawing.Size(140, 13);
+            this.label2.Size = new System.Drawing.Size(254, 24);
             this.label2.TabIndex = 15;
             this.label2.Text = "Friendly name (for the regex)";
             // 
@@ -591,9 +610,10 @@
             // 
             this.preview.AutoSize = true;
             this.preview.ForeColor = System.Drawing.Color.Blue;
-            this.preview.Location = new System.Drawing.Point(13, 145);
+            this.preview.Location = new System.Drawing.Point(26, 290);
+            this.preview.Margin = new System.Windows.Forms.Padding(6, 0, 6, 0);
             this.preview.Name = "preview";
-            this.preview.Size = new System.Drawing.Size(77, 17);
+            this.preview.Size = new System.Drawing.Size(150, 31);
             this.preview.TabIndex = 17;
             this.preview.Text = "Preview: ...";
             // 
@@ -602,10 +622,11 @@
             this.combo.DrawMode = System.Windows.Forms.DrawMode.OwnerDrawFixed;
             this.combo.FormattingEnabled = true;
             this.combo.ItemHeight = 16;
-            this.combo.Location = new System.Drawing.Point(68, 19);
+            this.combo.Location = new System.Drawing.Point(136, 38);
+            this.combo.Margin = new System.Windows.Forms.Padding(6);
             this.combo.MaxDropDownItems = 10;
             this.combo.Name = "combo";
-            this.combo.Size = new System.Drawing.Size(408, 22);
+            this.combo.Size = new System.Drawing.Size(814, 22);
             this.combo.TabIndex = 0;
             this.combo.DrawItem += new System.Windows.Forms.DrawItemEventHandler(this.combo_DrawItem);
             this.combo.SelectedIndexChanged += new System.EventHandler(this.combo_SelectedIndexChanged);
@@ -619,9 +640,10 @@
             this.markAsNewEntry.CheckState = System.Windows.Forms.CheckState.Checked;
             this.markAsNewEntry.Enabled = false;
             this.markAsNewEntry.Font = new System.Drawing.Font("Microsoft Sans Serif", 7F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-            this.markAsNewEntry.Location = new System.Drawing.Point(241, 45);
+            this.markAsNewEntry.Location = new System.Drawing.Point(482, 90);
+            this.markAsNewEntry.Margin = new System.Windows.Forms.Padding(6);
             this.markAsNewEntry.Name = "markAsNewEntry";
-            this.markAsNewEntry.Size = new System.Drawing.Size(240, 17);
+            this.markAsNewEntry.Size = new System.Drawing.Size(422, 28);
             this.markAsNewEntry.TabIndex = 2;
             this.markAsNewEntry.Text = "Mark this as a new Entry in the Search History";
             this.markAsNewEntry.UseVisualStyleBackColor = true;
@@ -629,20 +651,61 @@
             // allColumns
             // 
             this.allColumns.AutoSize = true;
-            this.allColumns.Location = new System.Drawing.Point(68, 86);
+            this.allColumns.Location = new System.Drawing.Point(136, 172);
+            this.allColumns.Margin = new System.Windows.Forms.Padding(6);
             this.allColumns.Name = "allColumns";
-            this.allColumns.Size = new System.Drawing.Size(144, 21);
+            this.allColumns.Size = new System.Drawing.Size(274, 35);
             this.allColumns.TabIndex = 18;
             this.allColumns.Text = "Seach All Columns";
             this.allColumns.UseVisualStyleBackColor = true;
             // 
+            // toFilter
+            // 
+            this.toFilter.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
+            this.toFilter.Location = new System.Drawing.Point(1062, 336);
+            this.toFilter.Margin = new System.Windows.Forms.Padding(6);
+            this.toFilter.Name = "toFilter";
+            this.toFilter.Size = new System.Drawing.Size(150, 50);
+            this.toFilter.TabIndex = 19;
+            this.toFilter.Text = "To Filter";
+            this.toFilter.UseVisualStyleBackColor = true;
+            this.toFilter.Click += new System.EventHandler(this.toFilter_Click);
+            // 
+            // filterHelp
+            // 
+            this.filterHelp.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
+            this.filterHelp.AutoSize = true;
+            this.filterHelp.Font = new System.Drawing.Font("Microsoft Sans Serif", 8F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.filterHelp.Location = new System.Drawing.Point(765, 352);
+            this.filterHelp.Name = "filterHelp";
+            this.filterHelp.Size = new System.Drawing.Size(285, 26);
+            this.filterHelp.TabIndex = 20;
+            this.filterHelp.TabStop = true;
+            this.filterHelp.Text = "What is a Log Wizard Filter?";
+            this.filterHelp.LinkClicked += new System.Windows.Forms.LinkLabelLinkClickedEventHandler(this.filterHelp_LinkClicked);
+            // 
+            // negate
+            // 
+            this.negate.AutoSize = true;
+            this.negate.Location = new System.Drawing.Point(909, 172);
+            this.negate.Margin = new System.Windows.Forms.Padding(6);
+            this.negate.Name = "negate";
+            this.negate.Size = new System.Drawing.Size(134, 35);
+            this.negate.TabIndex = 21;
+            this.negate.Text = "Negate";
+            this.negate.UseVisualStyleBackColor = true;
+            this.negate.CheckedChanged += new System.EventHandler(this.negate_CheckedChanged);
+            // 
             // search_form
             // 
             this.AcceptButton = this.ok;
-            this.AutoScaleDimensions = new System.Drawing.SizeF(96F, 96F);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(192F, 192F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Dpi;
             this.CancelButton = this.cancel;
-            this.ClientSize = new System.Drawing.Size(618, 353);
+            this.ClientSize = new System.Drawing.Size(1236, 758);
+            this.Controls.Add(this.negate);
+            this.Controls.Add(this.filterHelp);
+            this.Controls.Add(this.toFilter);
             this.Controls.Add(this.allColumns);
             this.Controls.Add(this.markAsNewEntry);
             this.Controls.Add(this.combo);
@@ -665,7 +728,7 @@
             this.Controls.Add(this.label1);
             this.Font = new System.Drawing.Font("Microsoft Sans Serif", 10F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.Icon = ((System.Drawing.Icon)(resources.GetObject("$this.Icon")));
-            this.Margin = new System.Windows.Forms.Padding(4);
+            this.Margin = new System.Windows.Forms.Padding(8);
             this.Name = "search_form";
             this.StartPosition = System.Windows.Forms.FormStartPosition.CenterParent;
             this.Text = "Search For...";
@@ -749,5 +812,8 @@
         private System.Windows.Forms.ComboBox combo;
         private System.Windows.Forms.CheckBox markAsNewEntry;
         private System.Windows.Forms.CheckBox allColumns;
+        private System.Windows.Forms.Button toFilter;
+        private System.Windows.Forms.LinkLabel filterHelp;
+        private System.Windows.Forms.CheckBox negate;
     }
 }